requirements.txt
appdirs==1.4.3
asn1crypto==0.24.0
astroid==2.4.2
Brlapi==0.6.6
CacheControl==0.12.6
certifi==2019.11.28
chardet==3.0.4
colorama==0.4.3
command-not-found==0.3
contextlib2==0.6.0
cryptography==2.1.4
defer==1.0.6
defusedxml==0.6.0
distlib==0.3.0
distro==1.4.0
Django==2.2.13
django-allauth==0.42.0
django-cleanup==5.0.0
django-cors-headers==3.5.0
django-etc==1.2.0
django-filebrowser==3.13.1
django-grappelli==2.14.2
django-hitcount==1.3.3
django-js-asset==1.2.2
django-remote-forms==0.0.1
django-rest-auth==0.9.5
django-tinymce==3.0.2
djangorestframework==3.11.1
html5lib==1.0.1
httplib2==0.9.2
idna==2.8
importlib-metadata==1.7.0
ipaddr==2.2.0
isort==4.3.21
keyring==10.6.0
keyrings.alt==3.0
language-selector==0.1
lazy-object-proxy==1.4.3
lockfile==0.12.2
louis==3.5.0
mccabe==0.6.1
msgpack==0.6.2
netifaces==0.10.4
oauthlib==3.1.0
olefile==0.45.1
onboard==1.4.1
packaging==20.3
pep517==0.8.2
pexpect==4.2.1
Pillow==7.2.0
progress==1.5
psycopg2-binary==2.8.5
pycairo==1.16.2
pycrypto==2.6.1
pygobject==3.26.1
pylint==2.5.3
pyparsing==2.4.6
python-apt===1.6.4-elementary6-ubuntu5.0.1
python-decouple==3.3
python3-openid==3.2.0
pytoml==0.1.21
pytz==2020.1
pyxdg==0.25
PyYAML==3.12
reportlab==3.4.0
requests==2.22.0
requests-oauthlib==1.3.0
requests-unixsocket==0.1.5
retrying==1.3.3
SecretStorage==2.3.1
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
typed-ast==1.4.1
ubuntu-drivers-common==0.0.0
ufw==0.36
urllib3==1.25.8
virtualenv==15.1.0
webencodings==0.5.1
wrapt==1.12.1
xkit==0.0.0
zipp==3.1.0
